Moved RedBlack to separate files
